chore(journals-dtos): rename
